requirements.txt
numpy
pandas
matplotlib
selenium==4.23.1

torch
torchvision
transformers>=4.34.0
sentencepiece==0.1.99

streamlit==1.35.0
streamlit_lottie==0.0.5
streamlit_drawable_canvas==0.9.3
streamlit_antd_components==0.3.2
streamlit-extras==0.4.2
streamlit-mermaid==0.2.0

SQLAlchemy==2.0.25

huggingface-hub
modelscope
psutil==5.9.8
bitsandbytes>=0.39.0
accelerate>=0.24.1
omegaconf
python-multipart

openai>=1.34.0
einops>=0.7.0
sse-starlette>=2.1.0

langchain[all]
langchain-community
langchain-elasticsearch==0.2.2
langchain-openai
sentence-transformers
auto_gptq==0.7.1

faiss-gpu

audio_recorder_streamlit==0.0.8
pydub==0.25.1

albumentations==1.3.0
fvcore==0.1.5.post20221221
open_clip_torch==2.17.1
opencv_contrib_python
opencv_python==4.6.0.66
opencv_python_headless==4.6.0.66
Pillow==9.4.0
pytorch_lightning==1.5.0
safetensors==0.4.3
setuptools==66.0.0
share==1.0.4
submitit==1.5.1
torchmetrics==0.6.0
tqdm==4.66.4
